changed return type back to double

+++ include/crpropa/Grid.h
@@ -57,7 +57,7 @@ inline void reflectiveClamp(double x, int n, int &lo, int &hi, double &res) {
 }
 
 /** Symmetrical round */
-inline int round(double r) {
+inline double round(double r) {
 	return (r > 0.0) ? floor(r + 0.5) : ceil(r - 0.5);
 }
